Refactor simulation module: rename `step` to `steps`.
Renamed the `step` module to `steps` across the simulation directory to standardize naming and improve clarity. Updated all relevant imports, references, and documentation to reflect the new module name. Enhanced formatting in `fmt` implementations for better output readability.
