+++ src/simulation/walk.rs
@@ -4,7 +4,7 @@
    Date: 22/10/24
 ******************************************************************************/
 
-use crate::simulation::step::{Step};
+use crate::simulation::steps::{Step};
 use crate::Positive;
 use rust_decimal::Decimal;
 use std::fmt::Display;
